fileutil: Add initial directory-relative API
Modern UNIX come with a variety of filesystem API suffixed with "at",
like openat(), linkat(), etc.  The reason for their existence
is multiple.

First, if you're doing a lot of file operations in a directory, it's
simply more efficient to avoid having the kernel traverse potentially
long pathnames constantly.

Second, this avoids a problem where if a user does e.g.:
rm -rf somedir &
mv somedir othername
mkdir somedir
touch somedir/otherfile

We won't end up deleting otherfile, because all of our operations are
relative to the original "somedir".

The second rationale is unlikely to matter for OSTree since we'll
assume the user isn't going to move a repository around while we're
committing to it, but anyways, it's just better to use fd-relative
pathnames.